Add files via upload
The flag 'rst_inname' for WRF restart run does not work properly (at least for WRF4.2.1). To circumvent this problem, this Python script will read the folder for the restart files from 'rst_outname' in namelist.input and will set links to the restart files in the directory where WRF is started from. It needs to be run from the directory where WRF is started from as "python3 link_restart_files.py"